requirements.txt
python-dotenv==0.17.1
tweepy==3.10.0